refac: organize component to seperate concern
