+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export interface Message {
+    type: 'success' | 'error';
+    text: string;
+}
